java/src/test/java/com/yourtion/leetcode/daily/m12/d13/SolutionTest.java
package com.yourtion.leetcode.daily.m12.d13;

import com.yourtion.leetcode.TestUtils;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("每日一题 - 20201213")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments("[1,2,3,1]", true),
                arguments("[1,2,3,4]", false),
                arguments("[1,1,1,3,3,4,3,2,4,2]", true)
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void containsDuplicate(String source, boolean res) {
        System.out.printf("runTest: %s , res: %b", source, res);
        int[] ss = TestUtils.stringToIntegerArray(source);
        Assertions.assertEquals(res, new Solution().containsDuplicate(ss));
    }
}